+++ akshare/economic/macro_euro.py
@@ -42,6 +42,8 @@ def macro_euro_gdp_yoy():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(%)"]
     temp_df.name = "gdp_yoy"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -66,6 +68,8 @@ def macro_euro_cpi_mom():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(%)"]
     temp_df.name = "cpi_mom"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -90,6 +94,8 @@ def macro_euro_cpi_yoy():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(%)"]
     temp_df.name = "cpi_yoy"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -114,6 +120,8 @@ def macro_euro_ppi_mom():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(%)"]
     temp_df.name = "ppi_mom"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -138,6 +146,7 @@ def macro_euro_retail_sales_mom():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(%)"]
     temp_df.name = "retail_sales_mom"
+    temp_df = temp_df.astype(float)
     return temp_df
 
 
@@ -162,6 +171,8 @@ def macro_euro_employment_change_qoq():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(%)"]
     temp_df.name = "employment_change_qoq"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -186,6 +197,8 @@ def macro_euro_unemployment_rate_mom():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(%)"]
     temp_df.name = "unemployment_rate_mom"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -210,6 +223,8 @@ def macro_euro_trade_balance():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(亿欧元)"]
     temp_df.name = "trade_balance"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -234,6 +249,8 @@ def macro_euro_current_account_mom():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(亿欧元)"]
     temp_df.name = "current_account_mom"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -258,6 +275,7 @@ def macro_euro_industrial_production_mom():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值(%)"]
     temp_df.name = "industrial_production_mom"
+    temp_df = temp_df.astype(float)
     return temp_df
 
 
@@ -282,6 +300,8 @@ def macro_euro_manufacturing_pmi():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值"]
     temp_df.name = "manufacturing_pmi"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -306,6 +326,8 @@ def macro_euro_services_pmi():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值"]
     temp_df.name = "services_pmi"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -330,6 +352,8 @@ def macro_euro_zew_economic_sentiment():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值"]
     temp_df.name = "zew_economic_sentiment"
+    temp_df = temp_df.astype(float)
+    temp_df = temp_df[temp_df != 0]
     return temp_df
 
 
@@ -354,7 +378,61 @@ def macro_euro_sentix_investor_confidence():
     value_df.index = pd.to_datetime(date_list)
     temp_df = value_df["今值"]
     temp_df.name = "sentix_investor_confidence"
+    temp_df = temp_df.astype(float)
     return temp_df
+
+
+# 金十数据中心-伦敦金属交易所(LME)-持仓报告
+def macro_euro_lme_holding():
+    """
+    伦敦金属交易所(LME)-持仓报告, 数据区间从 20151022-至今
+    https://datacenter.jin10.com/reportType/dc_lme_traders_report
+    https://cdn.jin10.com/data_center/reports/lme_position.json?_=1591533934658
+    :return: 伦敦金属交易所(LME)-持仓报告
+    :rtype: pandas.DataFrame
+    """
+    t = time.time()
+    params = {
+        "_": str(int(round(t * 1000)))
+    }
+    r = requests.get("https://cdn.jin10.com/data_center/reports/lme_position.json", params=params)
+    json_data = r.json()
+    temp_df = pd.DataFrame(json_data["values"]).T
+    temp_df.fillna(value="[0, 0, 0]", inplace=True)
+    big_df = pd.DataFrame()
+    for item in temp_df.columns:
+        for i in range(3):
+            inner_temp_df = temp_df.loc[:, item].apply(lambda x: eval(str(x))[i])
+            inner_temp_df.name = inner_temp_df.name + "-" + json_data["keys"][i]["name"]
+            big_df = pd.concat([big_df, inner_temp_df], axis=1)
+    big_df.sort_index(inplace=True)
+    return big_df
+
+
+# 金十数据中心-伦敦金属交易所(LME)-库存报告
+def macro_euro_lme_stock():
+    """
+    伦敦金属交易所(LME)-库存报告, 数据区间从 20140702-至今
+    https://datacenter.jin10.com/reportType/dc_lme_report
+    https://cdn.jin10.com/data_center/reports/lme_stock.json?_=1591535304783
+    :return: 伦敦金属交易所(LME)-库存报告
+    :rtype: pandas.DataFrame
+    """
+    t = time.time()
+    params = {
+        "_": str(int(round(t * 1000)))
+    }
+    r = requests.get("https://cdn.jin10.com/data_center/reports/lme_stock.json", params=params)
+    json_data = r.json()
+    temp_df = pd.DataFrame(json_data["values"]).T
+    big_df = pd.DataFrame()
+    for item in temp_df.columns:
+        for i in range(3):
+            inner_temp_df = temp_df.loc[:, item].apply(lambda x: eval(str(x))[i])
+            inner_temp_df.name = inner_temp_df.name + "-" + json_data["keys"][i]["name"]
+            big_df = pd.concat([big_df, inner_temp_df], axis=1)
+    big_df.sort_index(inplace=True)
+    return big_df
 
 
 if __name__ == "__main__":
@@ -409,3 +487,10 @@ if __name__ == "__main__":
     # 金十数据中心-经济指标-欧元区-领先指标-欧元区Sentix投资者信心指数报告
     macro_euro_sentix_investor_confidence_df = macro_euro_sentix_investor_confidence()
     print(macro_euro_sentix_investor_confidence_df)
+
+    # 金十数据中心-伦敦金属交易所(LME)-持仓报告
+    macro_euro_lme_holding_df = macro_euro_lme_holding()
+    print(macro_euro_lme_holding_df)
+    # 金十数据中心-伦敦金属交易所(LME)-库存报告
+    macro_euro_lme_stock_df = macro_euro_lme_stock()
+    print(macro_euro_lme_stock_df)