+++ akshare/stock/zh_stock_a_sina.py
@@ -39,7 +39,7 @@ def _get_zh_a_page_count() -> int:
 
 def stock_zh_a_spot() -> pd.DataFrame:
     """
-    从新浪财经-A股获取所有A股的实时行情数据, 大量抓取容易封IP
+    从新浪财经-A股获取所有A股的实时行情数据, 重复运行本函数会被新浪暂时封 IP
     http://vip.stock.finance.sina.com.cn/mkt/#qbgg_hk
     :return: pandas.DataFrame
                 symbol    code  name   trade pricechange changepercent     buy  \
@@ -84,15 +84,32 @@ def stock_zh_a_spot() -> pd.DataFrame:
     zh_sina_stock_payload_copy = zh_sina_a_stock_payload.copy()
     for page in tqdm(range(1, page_count+1), desc="Please wait for a moment"):
         zh_sina_stock_payload_copy.update({"page": page})
-        res = requests.get(
+        r = requests.get(
             zh_sina_a_stock_url,
             params=zh_sina_stock_payload_copy)
-        data_json = demjson.decode(res.text)
+        data_json = demjson.decode(r.text)
         big_df = big_df.append(pd.DataFrame(data_json), ignore_index=True)
+    big_df = big_df.astype({"trade": "float",
+                            "pricechange": "float",
+                            "changepercent": "float",
+                            "buy": "float",
+                            "sell": "float",
+                            "settlement": "float",
+                            "open": "float",
+                            "high": "float",
+                            "low": "float",
+                            "volume": "float",
+                            "amount": "float",
+                            "per": "float",
+                            "pb": "float",
+                            "mktcap": "float",
+                            "nmc": "float",
+                            "turnoverratio": "float",
+                            })
     return big_df
 
 
-def stock_zh_a_daily(symbol: str = "sh600582", adjust: str = "") -> pd.DataFrame:
+def stock_zh_a_daily(symbol: str = "sz000613", adjust: str = "qfq") -> pd.DataFrame:
     """
     新浪财经-A股-个股的历史行情数据, 大量抓取容易封IP
     :param symbol: sh600000
@@ -118,7 +135,7 @@ def stock_zh_a_daily(symbol: str = "sh600582", adjust: str = "") -> pd.DataFrame
     amount_data_df = pd.DataFrame(amount_data_json)
     amount_data_df.index = pd.to_datetime(amount_data_df.date)
     del amount_data_df["date"]
-    temp_df = pd.merge(data_df, amount_data_df, left_index=True, right_index=True, how="left")
+    temp_df = pd.merge(data_df, amount_data_df, left_index=True, right_index=True, how="outer")
     temp_df.fillna(method="ffill", inplace=True)
     temp_df = temp_df.astype(float)
     temp_df["amount"] = temp_df["amount"] * 10000
@@ -188,7 +205,7 @@ def stock_zh_a_daily(symbol: str = "sh600582", adjust: str = "") -> pd.DataFrame
 if __name__ == "__main__":
     stock_zh_a_daily_hfq_df = stock_zh_a_daily(symbol="sh600582", adjust="qfq-factor")
     print(stock_zh_a_daily_hfq_df)
-    stock_zh_a_daily_df = stock_zh_a_daily(symbol="sh600582", adjust="qfq")
+    stock_zh_a_daily_df = stock_zh_a_daily(symbol="sz000613", adjust="qfq")
     print(stock_zh_a_daily_df)
-    # stock_zh_a_spot_df = stock_zh_a_spot()
-    # print(stock_zh_a_spot_df)
+    stock_zh_a_spot_df = stock_zh_a_spot()
+    print(stock_zh_a_spot_df)