+++ akshare/stock_fundamental/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding:utf-8 -*-
 # /usr/bin/env python
 """
-Date: 2019/9/30 13:58
-Desc:
+Date: 2020/5/14 15:34
+Desc: 
 """